+++ example/src/WebApi/PubSub/Validators/FooContractValidator.cs
@@ -1,0 +1,20 @@
+using FluentValidation;
+using HumanaEdge.Webcore.Core.DependencyInjection.Validators;
+
+namespace HumanaEdge.Webcore.Example.WebApi.PubSub.Validators
+{
+    /// <summary>
+    /// Here's some Foo.
+    /// </summary>
+    [Validator(typeof(FooContract))]
+    public class FooContractValidator : AbstractValidator<FooContract>
+    {
+        /// <summary>
+        /// ctor.
+        /// </summary>
+        public FooContractValidator()
+        {
+            RuleFor(x => x.Name).NotEmpty();
+        }
+    }
+}